Support composite module outputs; preserve dtype
Update Hook.get_full_hook_fn to handle module outputs that are tuples or lists (e.g., CLIP/MaPLe blocks) by extracting the activation tensor, applying the hook, and reinserting the hooked activations while preserving auxiliary state. Add backward compatibility for hooks that return a single-element tuple/list. Also ensure processed activations keep the original activation dtype in create_sae_hooks to avoid float/half mismatches in MaPLe attention layers.

+++ patchsae-main/tasks/classification_with_top_k_masking.py
@@ -87,7 +87,8 @@ def create_sae_hooks(vit_type, cfg, cls_features, sae, device, hook_type="on"):
             )[0]
             - SAE_BIAS
         )
-        return processed
+        # Keep original dtype to avoid Float/Half mismatches in MaPLe attention layers.
+        return processed.to(act.dtype)
 
     def hook_fn_default(activations):
         # In-place replacement of block activations for CLIP/base model.
